refactor: change the location of settings based screen to separate folder named simply 'settings'
