+++ app/src/main/java/com/devson/nvplayer/ui/screen/settings/RecycleBinScreen.kt
@@ -1,4 +1,4 @@
-package com.devson.nvplayer.ui.screen
+package com.devson.nvplayer.ui.screen.settings
 
 import android.app.Activity
 import android.net.Uri
