+++ app/src/main/java/com/devson/nvplayer/ui/screen/settings/ToolScreen.kt
@@ -1,4 +1,4 @@
-package com.devson.nvplayer.ui.screens.settings
+package com.devson.nvplayer.ui.screen.settings
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -6,7 +6,6 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.filled.AccessTime
 import androidx.compose.material.icons.filled.Search
-import androidx.compose.material.icons.filled.VideoSettings
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
